pyproject.toml
[project]
name = "youtube-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "loguru>=0.7.2",
    "openai-whisper>=20240930",
    "pydub>=0.25.1",
    "pytest>=8.3.3",
    "pytube>=15.0.0",
    "yt-dlp>=2024.11.4",
]

[tool.pytest.ini_options]
addopts = "-m 'not slow'"  # Skip slow tests by default
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]
